feat: update the "panel menu" extension point to use the hooks
