+++ public/app/features/dashboard/dashgrid/PanelHeader/PanelHeaderMenuProvider.tsx
@@ -1,6 +1,13 @@
-import { ReactElement, useEffect, useState } from 'react';
+import { ReactElement, useEffect, useMemo, useState } from 'react';
 
-import { LoadingState, PanelMenuItem } from '@grafana/data';
+import {
+  LoadingState,
+  PanelMenuItem,
+  PluginExtensionPanelContext,
+  PluginExtensionPoints,
+  getTimeZone,
+} from '@grafana/data';
+import { usePluginLinkExtensions } from '@grafana/runtime';
 import { getPanelStateForModel } from 'app/features/panel/state/selectors';
 import { useSelector } from 'app/types';
 
@@ -21,10 +28,36 @@ interface Props {
 export function PanelHeaderMenuProvider({ panel, dashboard, loadingState, children }: Props) {
   const [items, setItems] = useState<PanelMenuItem[]>([]);
   const angularComponent = useSelector((state) => getPanelStateForModel(state, panel)?.angularComponent);
+  const context = useMemo(() => createExtensionContext(panel, dashboard), [panel, dashboard]);
+  const { extensions } = usePluginLinkExtensions({
+    extensionPointId: PluginExtensionPoints.DashboardPanelMenu,
+    context,
+    limitPerPlugin: 3,
+  });
 
   useEffect(() => {
-    setItems(getPanelMenu(dashboard, panel, angularComponent));
-  }, [dashboard, panel, angularComponent, loadingState, setItems]);
+    setItems(getPanelMenu(dashboard, panel, extensions, angularComponent));
+  }, [dashboard, panel, angularComponent, loadingState, setItems, extensions]);
 
   return children({ items });
+}
+
+function createExtensionContext(panel: PanelModel, dashboard: DashboardModel): PluginExtensionPanelContext {
+  return {
+    id: panel.id,
+    pluginId: panel.type,
+    title: panel.title,
+    timeRange: dashboard.time,
+    timeZone: getTimeZone({
+      timeZone: dashboard.timezone,
+    }),
+    dashboard: {
+      uid: dashboard.uid,
+      title: dashboard.title,
+      tags: Array.from<string>(dashboard.tags),
+    },
+    targets: panel.targets,
+    scopedVars: panel.scopedVars,
+    data: panel.getQueryRunner().getLastResult(),
+  };
 }
